pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "abulafia"
version = "0.1.3"
description = "A tool for fair and reproducible crowdsourcing using Toloka"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Tuomo Hiippala"},
    {name = "Helmiina Hotti"}
]
dependencies = [
    "crowd_kit>=1.0.0",
    "pytest",
    "PyYAML>=6.0",
    "toloka_kit[pandas]>=1.0.0",
    "wasabi>=0.9.0"
]
